# docs/echo/openapi.yaml
openapi: 3.0.3
info:
  title: Gateway HTTP API
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/experiments/user:
    post:
      operationId: UserExperiment
      tags: [experiments]
      summary: Get user experiments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExperimentRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentsReply'
...
  /v1/feature-toggles:
    post:
      operationId: CreateFeatureToggle
      tags: [feature-toggles]
      summary: Create feature toggle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFeatureToggleRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockReply'
  /v1/feature-toggles/{feature_toggle_id}/rollout:
    patch:
      operationId: UpdateFeatureToggleRollout
      tags: [feature-toggles]
      summary: Update feature toggle rollout
      parameters:
        - name: feature_toggle_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFeatureToggleRolloutRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockReply'
  /v1/feature-toggles/{feature_toggle_id}/status:
    post:
      operationId: SetFeatureToggleStatus
      tags: [feature-toggles]
      summary: Set feature toggle status
      parameters:
        - name: feature_toggle_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetFeatureToggleStatusRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockReply'
...
  /v1/feature-toggles/user:
    post:
      operationId: IsUserInFeature
      tags: [feature-toggles]
      summary: Check user feature toggles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IsUserInFeatureRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IsUserInFeatureReply'
components:
  schemas:
    ErrInfoReason:
      type: string
      enum:
        - UNSPECIFIED
        - VALIDATION_ERROR
        - INVALID_REQUEST
        - STATUS_OK
    StockReply:
      type: object
      required:
        - err_info_reason
        - message
      properties:
        err_info_reason:
          $ref: '#/components/schemas/ErrInfoReason'
        message:
          type: string
...
    ExperimentRequest:
      type: object
      required:
        - split_id
        - device_id
        - namespace
      properties:
        split_id:
          type: integer
          format: int64
        device_id:
          type: integer
          format: int64
        namespace:
          type: string
        city:
          type: string
        store:
          type: string
        params:
          type: array
          items:
            $ref: '#/components/schemas/Param'
    Param:
      type: object
      required:
        - param_name
        - value
      properties:
        param_name:
          type: string
        value:
          type: string
    ExperimentsReply:
      type: object
      required:
        - err_info_reason
        - experiments_reply
      properties:
        err_info_reason:
          $ref: '#/components/schemas/ErrInfoReason'
        experiments_reply:
          type: array
          items:
            $ref: '#/components/schemas/ExperimentReply'
    ExperimentReply:
      type: object
      required:
        - experiment_name
        - group_name
      properties:
        experiment_name:
          type: string
        group_name:
          type: string
    FeatureToggleStatus:
      type: string
      enum:
        - draft
        - active
        - disabled
        - archived
    CreateFeatureToggleRequest:
      type: object
      required:
        - name
        - namespace_id
      properties:
        name:
          type: string
          example: test_feature_toggle
        namespace_id:
          type: integer
          format: int64
          example: 1
        ios_rollout_percentage:
          type: integer
          format: int64
          minimum: 0
          maximum: 100
          example: 5
        android_rollout_percentage:
          type: integer
          format: int64
          minimum: 0
          maximum: 100
          example: 10
        web_rollout_percentage:
          type: integer
          format: int64
          minimum: 0
          maximum: 100
          example: 15
        rollout_percentage:
          type: integer
          format: int64
          minimum: 0
          maximum: 100
          example: 5
    UpdateFeatureToggleRolloutRequest:
      type: object
      properties:
        ios_rollout_percentage:
          type: integer
          format: int64
          minimum: 0
          maximum: 100
          example: 5
        android_rollout_percentage:
          type: integer
          format: int64
          minimum: 0
          maximum: 100
          example: 10
        web_rollout_percentage:
          type: integer
          format: int64
          minimum: 0
          maximum: 100
          example: 15
        rollout_percentage:
          type: integer
          format: int64
          minimum: 0
          maximum: 100
          example: 5
    SetFeatureToggleStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/FeatureToggleStatus'
    IsUserInFeatureRequest:
      type: object
      required:
        - user_id
        - namespace
        - platform
      properties:
        user_id:
          type: integer
          format: int64
          example: 123
        namespace:
          type: string
          example: mobile_app
        platform:
          type: string
          example: ios
    IsUserInFeatureReply:
      type: object
      required:
        - err_info_reason
        - features
      properties:
        err_info_reason:
          $ref: '#/components/schemas/ErrInfoReason'
        features:
          type: array
          items:
            $ref: '#/components/schemas/Feature'
    Feature:
      type: object
      required:
        - feature_id
        - feature_name
      properties:
        feature_id:
          type: integer
          format: int64
          example: 1
        feature_name:
          type: string
          example: new_checkout